+++ dessia_common/datatools/dataset.py
@@ -12,7 +12,7 @@ from sklearn import preprocessing
 
 try:
     from plot_data.core import Scatter, Histogram, MultiplePlots, Tooltip, ParallelPlot, PointFamily, EdgeStyle, Axis, \
-        PointStyle
+        PointStyle, Sample
     from plot_data.colors import BLUE, GREY
 except ImportError:
     pass
@@ -703,8 +703,10 @@ class Dataset(DessiaObject):
 
     def _plot_data_list(self):
         plot_data_list = []
-        for row, _ in enumerate(self.dessia_objects):
-            plot_data_list.append({attr: self.matrix[row][col] for col, attr in enumerate(self.common_attributes)})
+        for row, dobject in enumerate(self.dessia_objects):
+            sample_values = {attr: self.matrix[row][col] for col, attr in enumerate(self.common_attributes)}
+            reference_path = f"dessia_objects/{row}"
+            plot_data_list.append(Sample(sample_values, reference_path, dobject.name))
         return plot_data_list
 
     def _point_families(self):